+++ calvin/tests/__init__.py
@@ -2,6 +2,8 @@
 from mock import Mock
 
 from calvin.runtime.north import metering
+from calvin.utilities import calvinuuid
+from calvin.runtime.north.fifo import FIFO
 
 
 class DummyNode:
@@ -15,3 +17,40 @@ class DummyNode:
 
     def calvinsys(self):
         return None
+
+
+class TestNode:
+
+    def __init__(self, uri):
+        self.id = calvinuuid.uuid("NODE")
+        self.uri = uri
+
+
+class TestActor:
+
+    def __init__(self, name, type, inports, outports):
+        self.id = calvinuuid.uuid("ACTOR")
+        self.name = name
+        self._type = type
+        self.inports = inports
+        self.outports = outports
+
+
+class TestPort:
+
+    def __init__(self, name, direction):
+        self.id = calvinuuid.uuid("PORT")
+        self.name = name
+        self.direction = direction
+        self.peer = None
+        self.peers = None
+        self.fifo = FIFO(5)
+
+    def is_connected(self):
+        return True
+
+    def get_peer(self):
+        return self.peer
+
+    def get_peers(self):
+        return self.peers
